Update `UserControllerSpec` to Javalin 4
I had to add several attributes to the `ContextUtil` mock context object used in the testing to get things to work in the new universe.

The big thing was that I had to inject was a `JavalinJackson` mapper that is now used to convert between POJOs and JSON objects. I also had to inject a `maxRequestSize` parameter that `ContextUtil` needs for all the POST requests.

Constructing a `ContextUtil` got complicated enough that I pulled that out into two helper functions.

Then every test that used to call `ContextUtil.init()` directly had to be changed to use the new setup.