Redesign unit test for get all users by age
This adds a new `getUsersByAge()` test that follows the pattern in the Javalin testing examples for more "purely" unit tests.

I brought in Mockito to mock the `Context` class and then spec'ed what the mock context would return when `queryParamMap()` and `queryParamAsClass()` were called.

These `when().thenReturn()` are more complicated that I'd like in part because by default Mockito won't allow you to mock chained calls. You *can* do it, but they strongly discourage it (https://www.javadoc.io/doc/org.mockito/mockito-core/2.13.0/org/mockito/Mockito.html#RETURNS_DEEP_STUBS), so I didn't want to immediately go there.

So in a lot of ways I'd rather have something like:

   when(ctx.queryParamMap().contains("age")).thenReturn(true)

To do that in a way that the Mockito people would like, I think we'd need to make a subclass of Context that has direct methods for things like seeing if there's an "age" key. Seems a bit heavy to me at the moment, but I could be talked into trying it.

Rather than capturing the argument to `ctx.json()`, I used the Mockito `ArgumentMatcher`, which is what the Mockito folks recommend that you use instead of argument capturing. That takes a function that checks the argument in question. In our case this means looping through the list and asserting that everyone has age 37.