Replace pattern in regex for ensuring case insensitivity
I think this solution takes the mystery out of what the "i" was about (like you said earlier might be better), and I'm wondering if we were missing a Pattern.compile method that was needed, and that was causing the problem. This explanation leaves me to believe that Pattern.quote was not returning a Pattern, and that might have been our problem. Incidentally, it might also explain why I didn't know where to look for where we compile the pattern... because we didn't.

Was there a failing test we could use to capture the described problem? This passes our checks, but I'm not sure I totally understand the bug, and I don't know if it would show up in e2e or is we compensate for the case (in)sensitivity elsewhere that would mask a problem if it still existed here.

Closes #43 (I think)