Update Cypress configuration and file structure
This includes replacing the configuration json file with a typescript file and moving the e2e tests to a folder called e2e. I believe that there was a time when Cypress was using a folder called `integration` by default (maybe to differentiate it from Protractor tests?), but now it seems to be e2e by default, so we will follow suit. 

Also, there is some problem we've had with Chai assertions and Jasmine assertions not playing nicely together and causing lots of errors to show up in VS Code. The change to tsconfig.json makes that problem go away.