Add dependency from Iteration Template
The one thing I need to add (aside from the docker stuff that we don't use here) from the Iteration Template to make this match what we have there. For some reason, putting this in for the previous update made the changes look much more drastic for the other commit about dependencies, so I held onto it for its own step.