Change to `UntypedFormGroup`
This change matches our Iteration Template and might be important since `FormGroup` doesn't mean what it used to now that `FormGroup` elements are strongly typed by default. Since I'm not updating using the CLI, this kind of thing could be missed easily. It might not be a problem, but just in case it is, I will make it match what would have happened with the CLI update.

See https://github.com/UMM-CSci-3601/3601-iteration-template/issues/977 for more information. In particular, I'm not sure if our `FormGroup` will work because I think it might need to have the same type for all of the elements, and ours does not. It's possible that means within the options for a single input. I'd have to read more carefully.