Update `user.service.spec.ts`(to match a branch in Lab 3)
In this case, I actually pulled the Karma tests from the branch in lab 3 where we are working to modernize the Karma tests. There was such nice work there that I could not let it go to waste. I also added a test for adding users. So, this doesn't match the Iteration Template at this point.