Update `UserProfileComponent` in a recommended way
Much of the work done in this commit is modeled after a suggestion in this StackOverflow answer: https://stackoverflow.com/questions/38008334/angular-rxjs-when-should-i-unsubscribe-from-subscription#:~:text=In%20most%20cases%20we%20will,error()%20messages%20are%20published

This set of changes not only makes sure that subscription is handled in a way that completes any observables, but also means we don't need to (somewhat artificially) trigger any updates in the component that allow us to poke around in the paramMap and get new user information.

We also add the MatSnackBar for giving information to the user of the webpage if loading the user doesn't work instead of just putting that information in the console.

Co-authored-by: Nic McPhee <mcphee@morris.umn.edu>