Incorporate a way of testing by age that was explored by Nic
Nic made a branch and PR exploring a different way to do the Javalin tests. Some of this was addressed in a major overhaul of how the Javalin testing was done, but there are some things Nic tried in this branch that seemed like they might be helpful to retain.

This commit also involves removing some magic numbers by replacing them with constants. Those constants need to be available to the package for them to be used in these tests.