Group users by company
This adds a new `usersByCompany` endpoint which returns the users (just ID and name) grouped by company.

I had to create two new Java classes to capture the structure of these results. `UserGroupResult` (which should be renamed) holds the company (`_id`), the `count` (how many users work for that company), and `users` (a list of `UserIdName` objects). The `UserIdName` objects just contain the `_id` and `name` of a `User`.

GitHub CoPilot was a huge help here, both in helping generate the queries, but also in debugging some `IOException` errors that were the result of mismatches between the structures being generated by the queries and my Java classes.